Refactoring: Now using Java's Properties class for application configuration
